pyproject.toml
[tool.poetry]
name = "python_template"
version = "0.1.0"
description = "A simple starter for a python project"
authors = ["unruffled-nightingale <unruffled.nightingale@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
alive-progress = "~2.4.1"
beautifulsoup4 = "~4.11.1"
inquirer = "~2.10.0"
requests = "~2.28.1"
rich = "~12.6.0"
SQLAlchemy = "~1.4.41"
spotipy = "~2.21.0"
python-dotenv = "~0.21.0"
types-requests = "^2.28.11"


[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
flake8 = "*"
flake8-pyproject="*"
isort = "*"
mypy = "*"
pre-commit = "*"
pre-commit-hooks = "*"
pytest = "*"
pytest-cov = "*"
pytest-randomly = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/options.html
line_length = 120
use_parentheses = "True"
profile = "black"


[tool.flake8]
max-line-length = 120
exclude = "venv,.venv,.git,__pycache__,build,dist"


[tool.pytest.ini_options]
# https://flake8.pycqa.org/en/latest/user/options.html#configuration-options
# https://docs.pytest.org/en/6.2.x/reference.html#command-line-flags
addopts = "-ra -v"

[tool.coverage.report]
# https://coverage.readthedocs.io/en/latest/config.html#report
show_missing = true

[tool.coverage.run]
# https://coverage.readthedocs.io/en/latest/config.html#run
omit = ["./tests/*"]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
exclude = "tests"
disallow_any_unimported = "False"
disallow_any_generics = "False"
disallow_incomplete_defs = "True"
disallow_subclassing_any = "False"
disallow_untyped_calls = "True"
disallow_untyped_defs = "True"
disallow_untyped_decorators = "True"
ignore_missing_imports = "True"
no_implicit_optional = "True"
pretty = "True"
show_error_codes = "True"
show_error_context = "True"
strict_equality = "True"
strict = "True"
warn_return_any = "False"
warn_unused_ignores = "True"
warn_unreachable = "True"
warn_redundant_casts = "True"
warn_unused_configs = "True"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_any_unimported = "False"
disallow_untyped_calls = "False"
no_implicit_optional = "True"